Replace custom IdlingResource with RxIdler
Our custom IdlingResource used RxJavaPlugins.setScheduleHandler() to
keep track of any active tasks. The problem with this approach was that
a task was only considered active once it actually started executing,
rather than as soon as it was scheduled. This caused test flakiness, as
there could be an arbitrarily long delay between scheduling and
execution (due to the overhead of starting a new thread or a thread pool
being saturated, for example).

RxIdler solves this problem through a delegating Scheduler that
considers a task active as soon as it is scheduled.